Don't call ComputeIntermediateSatelliteAssemblies
Calling this target directly can cause it to run out of order.  This was
happening when the pkg project built before source, and was causing
CreateManifestResourceNames to remove all EmbeddedResources here
https://github.com/Microsoft/msbuild/blob/d42d3504057ef2b88dd4f68c4bfc5591371bd6fe/src/Tasks/Microsoft.CSharp.CurrentVersion.targets#L121
because we hadn't run the targets that set %(EmbeddedResource.Type).

Instead of doing this depend on the BuildOutputGroup.
Output groups are meant to be safe to call in isolation, so they should
always fully specify their target dependency chain to produce the
correct items without modifying the project state.