Fix misport where failure occurs in Release config
A single line was left outside of a `#if DEBUG` directive block, causing
it to have an invalid state Release config builds. It looks like this
was just a simple misport because moving the line back under DEBUG
introduces an uninitialized field warning. Setting the field to null in its
initializer is the proper fix for that warning.