Streamline local dev configuration
Add a PowerShell script that simplifies and streamlines the UX around
installing a new version of SDK and getting VS to pick it up.

Our local build scripts download and install a required version of SDK
into a local folder (.\.dotnet\sdk). This is done to accommodate build
infrastructure requirement.

VS is only loading SDKs from under C:\Program Files\dotnet\sdk folder.
This folder is also protected and requires an elevation to write into it.

The script can be run as follows:

    PS> .\build-local.ps1                   # to simply build and launch the solution in VS
    PS> .\build-local.ps1 -FullClean        # resets the SDK and the local dev environment

The new script does a number of things:

  - Checks whether the version of SDK required in global.json is installed
both locally and system-wide. If it isn't the script is re-run in an
elevated mode to establish a symlink from the local SDK copy to the
system-wide;
  - If a full cleanup is required,
      * kills all running instances of dotnet.exe and MSBuild.exe, because
    they can lock SDK files and prevent removal of the SDK folders,
      * deletes the system-wide SDK installation, and
      * cleans the local repository;
  - Kicks off `build.cmd` to downloads the required SDK, build the
     solution and run tests;
  - Restores necessary packages;
  - Launches the solution in Visual Studio

After running the script once, the solution can now be opened in VS by
simply clicking on the solution file in Windows Explorer.