Fix tooltip crash and harden
The tooltip class will cause Windows to AV when attempting to get the tooltip bubble size when there isn't an actively showing tooltip. (Really Windows should be checking before derefing their internal pointer, but this has been there forever.) Fixing this by making sure to check that one is active when sending the query message.

The existing TooInfo usage was not safe and could crash with bad timing with a GC. I've hardened usage of the TTTOOLINFO struct and made it blittable, which should make things both safe and more performant as we're not instantiating a class anymore.

Fixes #1602.