I had the interfaces wrong and didn't notice until I tried using IPicutreDisp. It was incorrect before.
`OLE_HANDLE` is a uint, no matter what the platform. The dispatch invocation throws as it sees that the native type is uint, which doesn't match IntPtr. Invoking through the IUnknown RCW doesn't have any type info to check so it works fine due to the luck of calling conventions matching up.

The IPictureDisp version never worked on 64 bit. Now it does. :)