Properly add and remove reflect children
Fixes #1311
Fixes #1520
Fixes #1607
Fixes #1644

The bug was introduced as part of refactoring in #1034 - ordering got flipped with switching `as` to `is`.
The confusion was likely caused by the fact that the property setter called its own getter.

Update the code to be more explicit in terminology to hopefully avoid this from happening again.