fix PowerStatus_BatteryChargeStatus_Get_ReturnsExpected test
PowerStatus_BatteryChargeStatus_Get_ReturnsExpected was not correctly
validating combination of flags.
The test was likely written on a workstation permanently connected to a
power point. As a result the test did not account for combination of
flags that can occur on laptops that are not permanently connected,
e.g. if a laptop was being charged querying power would result in a
combination of flags, such as `HIGH | CHARGING`.

Enhance the test to assert on combinations of valid flags.