Move NativeWindow imports to interop
Move a number of the imports used in NativeWindow to the new interop pattern. Also introduce WindowMessage to match the approach in the designer.

The idea with WindowMessage is to make it signficantly easier to debug.  It is extremely mentally taxing to follow message pumping with `int`.  Converting over WindowMessages is a huge change, so doing it in phases.