fix `COM2PropertyDescriptor.SetValue` incorrect order of precedence
`COM2PropertyDescriptor.SetValue` had an incorrect order of precedence
that would lead to IOORE if error info string was empty.

Extract the code in question in a standalone method so it could be tested.
After the tests, the implementation is rewritten to reduce allocations.

An alternative and significantly shorter implementation is considered:
`strMessage.ToString().TrimEnd('\r', '\n')` however uses significantly
more memory, though it is slightly more performant.

```
|      Method |       Mean |      Error |     StdDev | Ratio |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------ |-----------:|-----------:|-----------:|------:|-------:|-------:|------:|----------:|
|     Iterate | 1,923.3 ns | 13.6586 ns | 12.1080 ns |  1.00 | 3.0899 | 0.0076 |     - |   19384 B |
| TrimNewline |   207.7 ns |  0.8548 ns |  0.7995 ns |  0.11 | 0.0725 |      - |     - |     456 B |
|  TrimString |   165.8 ns |  1.3424 ns |  1.1900 ns |  0.09 | 0.1631 | 0.0005 |     - |    1024 B |
```