APIcompat
Proposed changes
* share resources when possible with Gen Facades target
* small cleanup of some targets files related to Gen Facades
* plumb API Compat into build, test against 4.7.2 Framework assemblies
* add baseline files for thorough review

Uses APICompat via Arcade to verify implementation in build against the
references in microsoft.targetingpack.netframework.v4.7.2.

1. `Tools.props` - responsible for downloading the 4.7.2 targeting pack
that will be used to resolve matching contract. This props file is
automatically used by Arcade if it is present.

2. `Directory.Build.props` - brings in `DesktopAPICompat.props` to each
dll project; also establishes the path to the Desktop References
downloaded by Tools.props.

3. `ResolveMatchingContract.targets` - responsible for determining if
there is a matching dll in the 4.7.2 targeting pack for each of our assemblies.
This is true for the `System.Design`, `System.Drawing`, and
`System.Drawing.Design` facades as well as `System.Windows.Forms` and
`Accessibility`; matching is performed by `AssemblyName`, and this target
will throw an error if the path to the Desktop References established in
`Directory.Build.props` does not exist to prevent a false positive in the
resolution of matching contacts. Uses the non-existence of a matching
contract to exclude projects which do not have a matching contract from
undergoing API Compat.
This target is shared by both `GenFacades` and `APICompat`.

4. `DesktopAPICompat.props` - establishes properties that will tell
APICompat to run for each project. Sets flags for the ApiCompatBaseline.txt
files. Also passes the list of attribute types to exclude from appearing
in baseline files (DesktopAttributeTypeExclusions.txt).

5. DesktopAPICompat.targets - includes the APICompat package reference
to projects which have been determined by `DesktopAPICompat.props` minus
`ResolveMatchingContract.targets` that they should undergo API compat.

6. `DesktopAttributeTypeExclusions.txt` - new-line delim'd list of
attributes to exclude addition/removal of from showing up as an error in
API Compat or having to be baselined. For example,
`System.Security.SecurityCriticalAttribute` was removed from .NET Core.

7. `ApiCompatBaseline.txt` files - establish the known changes between
the current implementation and the reference to desktop assemblies.
Changes which do not appear here will show up as errors in the build.
To regenerate these files, pass the `/p:BaselineAllAPICompatError=true`
flag to the build.

Resolves #1538
Resolves #1059