Move IHandle to Interop. (#2348)
This allows using the interface in other assemblies outside of System.Windows.Forms. We can't, however, use classes that implement from different assemblies. If we wanted to use Control in interop calls in assemblies other than System.Windows.Forms we'd have to make this an actual shared interface. The only practical way to do that is to introduce a higher level shared assembly that contains it. While SWF is effectively top-level, making internals visible would be less than ideal.

Ultimately we may want to create an Interop assembly that all of our other assemblies can rely on. It would make this really reusable, reduce code duplication, and generally make maintenance a bit easier.