fix: Unable to pack Microsoft.VisualBasic.Forms
Copy the project's intellisense file to the ref folder as a temporary fix
until the intellisense file will be added to the main docs bundle.
Without the file we were unable to pack.

Once the intellisense file is added to the main docs bundle, the task
will start breaking and remind us to revert the change.

Closes #2395