fix: Incorrect invocation of `PeekMessageW`
During the refactor in d52e526 a number of calls to `PeekMessageW` lost
`PM.REMOVE` argument that led to situations where callers would get
stuck in a infinite loop peeking for messages.

For example, a form with a `TabControl` with two `TabPage` controls added
to it would dead lock and hang.

Fix affected calls by restoring the correct argument.

Fixes #2462