Fix bad string.Format calls
These were identified by compiler warnings that appeared in the VS IDE.
I'm giving it a best guess as to which value was intended for inclusion as a formatting arg. Since these are all in a Debug.WriteLineIf call, I suspect they compile out of release builds so it was a low pri to fix.