Manual generation of transport package manifest
Make the process of updating the transport package manifest manual
to ensure changes in the manifest are made consciously.

During the `pack` command the exisintg manifest will be compared against
the list of assemblies in the nuspec file, and if those two differ -
the build will fail.

Addressing the feedback in https://github.com/dotnet/winforms/pull/2745#discussion_r370327755