fix: WebBrowser_AttachInterfaces_Invoke_Success fails if IE browser not installed
IE browser is now an optional Windows component, and it may not
be present in all environments we run tests.

`Type.GetTypeFromCLSID` returns `System.__ComObject` regardless of
whether the CLSID is valid. Which leaves us checking whether an IE
browser is installed and handling the exception.

Furhter xUnit does not support dynamic skipping of tests https://github.com/xunit/xunit/issues/2073
that forces us to abandon the test if the IE browser is not detected,
marking the test as success.

Relates to #2889 and dotnet/core-eng#9068