Remove unnecessary coherency attributes and prepare for strict… (#3113)
Remove the unnecessary coherency attributes. These coherency attributes
are not needed because the packages are produced in the same build as
the CPD parent.

This also prepares this repo for 'strict' coherency
https://github.com/dotnet/arcade/issues/5195. Strict coherency is
significantly simpler and resolves a few cases where legacy CPD can
downgrade dependencies in incremental servicing. It looks in the CPD
parent's version.details.xml file to find the the CPD dependency, then
updates to that version. In winforms case, the CPD's dependencies are not
dependended on within runtime's own build, and the CPD attribute should be removed.