Rework EventHandlerService tests
These tests aren't testing the public contract. Additionally, internals they are trying to manipulate aren't needed to validate the public contract.

It is important to not bind our tests any more tightly to internal implementation details than is necessary as it makes code changes (improvements / maintenance) much more difficult. It can also obscure what the public contract really is.

Internal access should be used to validate supporting methods and classes that aren't publicly visible. Setting internal fields to test methods should generally be avoided. There must be significant value in validating the public surface area to introduce this sort of coupling.

In this particular case there is not significant value in testing the way the caching is implemented, or even if there is caching. If performance is important, it should be validated via a performance test and/or tracking allocations for the thread. The internal linked list should have been factored into a collection type that could be tested on it's own. In this particular case there is a .NET collection that can be used that is equivalent enough.  I'll follow with that change.

We also should not be testing basic runtime functionality. Creating a new object never returns null without throwing, for example.