Fix menu arrow key navigation
Ignore the SystemInformation.RightAlignedMenus property when navigating
menus with the keyboard. From observation of the behavior of the
corresponding registry setting (MenuDropAlignment), it appears this
property is only intended to affect the display of the menus. Reversing
dependent on RightAlignedMenus caused the focus to shift in the wrong
direction when configured for left-handedness (the system default).