Correct ImageKey/ImageIndex properties
Setters for `ImageKey` and `ImageIndex` properties were made idempotent
in #3126, however the change didn't account for the fact that the default
`ImageKey == ""` and not `null`.

This caused failures to reset `ImageIndex`, when `ImageKey` was set to
`""`; and in reverse - failure to reset `ImageKey` when `ImageIndex` was
set to `-1`.

Tweak the setters to continue execution for default values.

In addition apply NRE annotations to ` ImageList.Indexer` type,
and expose the defaults in a reusable manner.

Resolves #3353