Implement CollapsedState to ListViewGroup (#3155)
Fixes #3067

ListViewGroup originally was not collapsible and thus items in the group
could not be temporarily hidden. These changes add a new property to
ListViewGroup which controls its appearance e.g. whether the group is
collapsible and if the group is in its collapsed state or expanded
state. These changes also raise an event for when the CollapsedState
of a group is changed.