fix: ContextMenuStrip leads to crash on W10 1607
W10 1607 is not fully compatible with PerMonv2, and this leads
to crashes interacting with ContextMenuStrip.

`CommonUnsafeNativeMethods.SetThreadDpiAwarenessContext` may return
`DPI_AWARENESS_CONTEXT_UNSPECIFIED` result that is never checked, however
we set `dpiAwarenessScopeIsSet` flag assuming the scope was set.

When disposing the scope we pass the obtained context to `TrySetThreadDpiAwarenessContext`
that tries to restore the scope back to the original. However in event
the original was `DPI_AWARENESS_CONTEXT_UNSPECIFIED` it leads to
ANE, that was observed in #2969.

Resolves #2696.