Fixing DPI changed events by changing the order in setting child Form owner (#3394)
* In a multi DPI environment, DPI changed events triggered only when there is a DPI change
happened for the Handle directly or via its parent. So, it is necessary to not set the
owner before creating the handle. Otherwise, the window may never receive DPI changed
event even if its parent has different DPI.Users at runtime, has to move the window
between the screens to get the DPI changed events triggered.

* Review feedback.