Correct SelectedImageKey/SelectedImageIndex/StateImageKey/StateImageIndex properties
Setters for `ImageKey` and `ImageIndex` properties were made idempotent
in #3126, however the change didn't account for the fact that the default
`ImageKey == ""` and not `null`.

This caused failures to reset `SelectedImageIndex` or `StateImageIndex`,
when `SelectedImageKey` or `StateImageKey` were set respectively to
`""`; and in reverse - failure to reset keys when indicies were
set to `-1`.

Tweak the setters to continue execution for default values.

Resolves #3425
Relates to #3364