Add scoping structs for Win32 objects
This builds on what I did for ScreenDC some time ago. It creates wrapper ref structs for deterministic undoing of resource creation and selection. The goal is to help prevent misordering and leaked objects while also reducing nesting complexity.

Note that discards for the new style usings is a coming enhancement in C# that will make this even cleaner. If we move towards typing native handles this pattern will work well with that.