Add region scoping
- Adds scope structs for GDI regions
- Removes WindowsRegion
- Rework WindowsGraphics and WindowsGraphicsWrapper to dispose safely/correctly
- Rework ActiveXImpl to not retain an active HRGN handle
- Fix some IntPtr != null checks

This is another step in trying to get rid of DeviceContext and WindowsGraphics and replace them with struct based contexts.

Remove WindowsGraphics finalizer in release builds

We only need the finalier for disposal tracking. DeviceContext has the actual native resource and will get finalized if we miss the disposal.

Removing these from the finalizer queue will improve performance and reduce risk.