Don't double apply Graphics transform
Grid entry text was being shifted out of the control window as we were applying the Graphics transform manually when the centralized code now does this for us.

This change no longer double applies and changes the code to use TextRenderer, removing the local HFONT caching.

Also fix GroupBox to use FillRect instead of Rectangle. Remove an uncalled overload for Rectangle.

Fixes #3580