Added UnitTests for System.Resources.ResXResourceSet
Needed to add four versions of RESX files to Unit Test project root directory because the files referenced in these RESX files will be looked for either relative to the RESX file (during development) OR relative to the binaries (during test runs), so the Unit Test project root directory is the only location consistent for both cases (development and test runs).FileStream.
The four versions of RESX files are for .NET Framework 1 (SWF 1), .NET Framework 2/3 (SWF 2), .NET Framework 4 (SWF 4) and current development version (SWF 42).
The unit tests cover loading of RESX file; retrieving and matching of data and meta-data strings as well as icon, bitmap and text (ansi and utf8) files stored or referenced in the RESX files (with resource names provided case sensitive and insensitive); and loading of correct Resource-Reader/Writer types.