Simplify HDC caching
We're getting intermittent errors in tests when trying to create a Graphics object from a cached DC. For 5.0 rolling this cache back to be per thread and no longer use the background thread to create HDCs. Also add additional diagonstics so we can better follow up on errors when they occur.