Do not reset user fonts in DataGridView
Whenever ambient font changed, this change would stomp over
user-defined DataGridView font styles, such as
* DefaultCellStyle
* ColumnHeadersDefaultCellStyle
* RowHeadersDefaultCellStyle

Whilst the behaviour goes all the way back to .NET Framework, it becomes
a significant migration issue for projects that want to keep the original
default font (that was changed in #656).

Prevent changes to either of the above cellstyles, if those configured by
a user.

Resolves #3033