Fix rendering of NumericUpDown control (#3939) (#3942)
GDI and GDI+ don't end their lines and rectangles on the same pixel. Lines end on the prior pixel and rectangles draw the right and bottom one pixel over.

- Fix the GDI coordinates to match the GDI+ rendering
- Remove the GDI+ fork as it is impossible to set a transparent background color for NumericUpDown
- Add initial rendering tests

Validated with A/B testing between 3.1 and 5.0 with a loud background color. Validated with and without visual styles, enabled and disabled, and with different border styles.