Apply clipping and translation transforms correctly
Logic was off for some of the cases where we set up an HDC for drawing and we weren't applying clipping or translation transforms when we needed to. For the ToolStrip this causes the background of the parent to fill over the ToolStrip background.

This has implications for most rendering where a clip or transform is applied.

Add tests for the HDC handling struct to validate application of Graphics properties.