Fix ListView is in virtual mode throws IOE #3962
During implementation of accessible objects for ListView we started to use SelectedItems property for getting of required items. Unfortunately, this property is not available when the ListView is in virtual mode. As result we got an exceptions when tried to do any UI action when ListView is in virtual mode.

As a fix, we implemented using SelectedIndices instead of SelectedItems when getting Accessible Objects, because SelectedIndices property is available in any mode.

Also we added a check for methods in ListViewGroup that the ListView is not in virtual mode, since ListViewGroups are not supported for virtual mode.

Besides we replaced "groups.Count > 0" condition on "GroupsEnabled" in logic for navigation on ListView by keyboard. Together with the "groups.Count > 0" condition, it also checks that Groups is displayed and ListView is in default (not virtual) mode