Fix "ListViewItem.ListViewItemAccessibleObject.Bounds" property
If a ListViewItem is in a collapsed ListViewGroup and we try to select it in Inspector then we get exception. It happens because during trying to get "Bounds" property we send a "GETITEMRECT" message to NativeControl and receive empty result.

To avoid this issue, we added the condition that if a ListViewItem is located in a ListViewGroup and this ListViewGroup is collapsed we return empty rectangle otherwise we call old logic for getting of rectangle