Fix "ListView.SelectedIndices.Contains" method
The "SelectedIndexies.Contains" method tries to get ListViewItem using the index from parameter and checks "Selected" property for this ListViewItem. If we close the dialog that contains ListView  then ListView is still available to us in the inspector, but ListViewItem will have index with "-1" value. As result if we call "SelectedIndexies.Contains" method for ListViewItem with such index then we get an exception.

As a fix we added check that the index should be greater than -1 and less than the number of Items in the ListView. If check is failed then we will return false, otherwise we will get the ListViewItem using index and check "Selected" property