Fix issue with announcing ExpandedCollapsedState for ComboBox in Property grid #1987
Issue is reproduced due to two competing requests to the Narrator. First request moves focus on selected item in options list, second request moves focus on control itself. As a result, in half of the cases the narrator announces the selected option, and in the other half the ExpandedCollapsedState of the control.

As a fix, I removed RaiseAutomationEvent and RaiseAutomationPropertyChangedEvent and replaced them on a RaiseAutomationNotification. Now, the focus always remains on the selected item and Narrator announces selected option and ExpandedCollapsedState of control.