Correct placeholder text drawing
Creation of a `Graphics` object is rather expensive. All paining should
also be done in `WM_PAINT` event, that's we we get a correct DC to
paint in. We should not paint in `WM_KILLFOCUS`.

The complexity comes from the fact that the native control tracks when
it needs repaining, and is unaware of our custom text we slap on top.
This causes a number of undesired side effects such as
- the native control stomps over our custom text, or
- we draw the text multiple times.

A user can also observe a very prominent flicker when a mouse is moved
in and out of the textbox while it show a placeholder text.
Dealign with the flicker is outside the scope of this change.

Co-authored-by: 		Jeremy Kuhne <JeremyKuhne@users.noreply.github.com>
Co-authored-by: 		Tobias Käs <weltkante@users.noreply.github.com>