Fix Application.EnableVisualStyles in single file publishing mode
Application.EnableVisualStyles gives Windows a manifest that enables the v6 common controls. We normally pull this manifest from a native resource in System.Windows.Forms.dll. In single file mode there is no "dll" to load from so this feature was not working as intended.

In the case where we can't get a Win32 HMODULE we now fall back to pulling the manifest from a managed resource and dumping it to the temp folder.

Make ACTCTXW blittable.