Fix `ListView` no longer displays images
Changes in `ImageList` ownership model in #3601 means that very are now two
instances of imagelists - one instance is tracked by Windows Forms (i.e. managed)
side, and another one tracked by the underlying Win32 (unmanaged) side.
This was done due to an observed ownership disconnect between the managed and unmanged
code, that led to situations where the unmanaged code would effectively destroy
and instance of an imagelist, which the managed code was oblivious to.

However with the above change changes to images in an imagelist on the managed
side, i.e. a user adding or replacing an image in the imagelist, would not be reflected
in the imagelist on the unmanaged side, and thus would not be reflected in the UI
(which is drawn by the Win32).

The fix reuses the established infrastructure that notifies the managed imagelist
implementation of changes to the images collection, and once a notification of
a chance is received, the unmanaged imagelist is re-created, thus ensuring the UI
has all the correct images to display.

Fixes #4169