Narrator and Inspect become unresponsive inspecting `Combobox` control (#4187)
Issue is reproduced due to incorrect implementation of the "ComboBoxAccessibleObject.IRawElementProviderHwndOverride" method. After adding of "ComImport" attribute, the "ComboBoxAccessibleObject.IRawElementProviderHwndOverride" method started to be used, which lead to an issue. As fix for it, we removed "ComImport" and stopped to using it.

I suggest to restore "ComImport" and remove overriding "ComboBoxAccessibleObject.IRawElementProviderHwndOverride" method.