Narrator and Inspect become unresponsive inspecting `Combobox` control
Issue is reproduced due to incorrect implementation of the
`ComboBoxAccessibleObject.IRawElementProviderHwndOverride` method.
After adding of `ComImport` attribute, the `IRawElementProviderHwndOverride`
method is called, which lead to an issue.
As fix for it, we removed "ComImport" and stopped to using it.

However the issue is not with `ComImport` but the override of
`ComboBoxAccessibleObject.IRawElementProviderHwndOverride` method.

Fixes #3452

(cherry picked from commit 2195c2c5687f695b1cdbf51db57af11eabe292d0)