Fix DataGridView Tip is not normal #2812
Added special logic that checks if the tooltip is internal/default or if tooltip uses a value from a cell. If the value is taken from a cell, we check if the value is cut off in the cell itself. If it is cut off, we display the tooltip with the value in the cell, otherwise the tooltip is not displayed. If tooltip is internal/default then we also display the tooltip