Fix issue with hanging of accessibility tools when the "ComboBox" to be expanded contains duplicate items (#4193)
Fixes #3590

Issue is reproduced because we used Combobox.Items.IndexOf() for getting of item index in ComboBoxItemAccessibleObject. If list contained "aaa" and "aaa" items then the "FragmentNavigate" method saw the first "aaa" item, tried to got the next item, but during getting data about second "aaa" method got the index of the first "aaa" item.  As result  the "FragmentNavigate" method thought it was seeing the first element and again tried to get next item. All this led to endless recursion.

As a fix, we reused the logic from the "ListBox", which uses "Entry" class instead of the strings to work with items. This avoids the problem when working with items with the same name.