Fix ComboBox does not respect "MaxDropDownItems" property #4373
The issue is reproduced because "MaxDropDownItems" property does not work when the "IntegralHeight" flag has "True" value. As a workaround, the user can update the "IntegralHeight" flag themselves for the usual ComboBox, but the user does not have access to this flag in the DataGridViewComboBoxColumn and DataGridViewComboBoxCell. As a result, this property becomes useless for the ComboBoxes in the DataGridView.

As a fix, I suggest to set "IntegralHeight" flag to "False" when the user updates "MaxDropDownItems" property of the DataGridViewComboBoxColumn or DataGridViewComboBoxCell