fix MDI clients lost control box icons
During a cleanup in #2659 the `GetNativeMenuItemImage` forgot to set the
size of the struct (previously done in the constructor when it was still
a class). Since the return code of `GetMenuItemInfoW` is not checked
it was not detected, effectively leaving the image at IntPtr.Zero and
ignoring the call.

Resolves #4388

(cherry picked from commit a40d23ce4227a25ae86b7f7bbaf9c7cff818fe5d)