Fix [Accessibility] An element of the given ControlType must support the Toggle pattern for PropertyGrid
In 4.7.2 these PropertyGrid buttons were of the "Checkbox" control type, which changed to "Button" from 4.8. After fixing #3618 we returned the behavior that was originally in 4.7.2, but we have lost support for the "Toggle" pattern. Since the buttons do not visually look like checkboxes, we wanted to change the "CheckBox" control type to "Button" control type that supports the "Toggle" pattern. Unfortunately, the "Accessibility Insights" rules prohibit the button to simultaneously support the "Invoke" and "Toggle" patterns. Therefore, based on the fact that the buttons have a behavior similar to radiobutons, as a fix, we decided to assign them a "RadioButton" control type.

Added logic to use "RadioButton" control type instead of "CheckBox" control type.

Added unit tests.