Fix Exception appears when Narrator is active and user tries to add row after removing of another row. #4228
The issue is reproduced because in some cases the "DataGridView" property may be null. Most of the methods and properties check that the "DataGridView" property is not null. But in some places (including the "FragmentRoot" property) we forgot to add this check and got an exception as a result.

Added skipped checks for null.

Added unit tests.