Fix The yellow rectangle is not displayed for a subitem when the user hovers over it or selects in the Inspector tree #4679
Changed the logic for getting ListViewSubItems for the Inspector tree.
Now they are not displayed for "List", "SmallIcon", "Large" views.
For "Tile" view, we display the ListViewSubItems in the Inspector tree based on the number of ListViewSubItems, the number of ListViewColumns and the "Bounds" property, through which we know that the ListViewSubItem is visible.
For "Details" view, we display ListViewSubItems based on the number of ListViewColumns. If the cell is visible and the ListViewSubItems does not exist, we create a fake ListViewSubItemsAccessibleObject for that cell.

Fixed the issue with getting a rectangle for ListViewSubItem when the ListView is in "Tile" view

Added missing conditions
Added unit and MAUI tests