Fix The Inspector displays incorrect rectangle for the ListViewGroup #4778
The issue with the incorrect rectangle is reproduced because the "FragmentRoot" property was not overridden. Added an override for the "FragmentRoot" property.

Fixed issue with getting a rectangle for a ListViewGroup. Now, instead of the ListViewGroup index, we use the group ID. Fixed a issue with getting an incorrect ListViewGroup index.

Added unit tests. Fixed typos in unit-tests naming