Fixed the bug when a ToolStripControlHost freezes UIA tools
(eg. Inspect) if they focus on a ToolStrip with the ToolStripControlHost.

The main reason - a hosted control, that is used as an item
of a ToolStrip via a ToolStripControlHost, doesn't provide info about parent
(the ToolStrip) and siblings (the ToolStrip next and previous items)

This commit fixes only controls which support UIA,
because these controls use FragmentRoot and FragmentNavigate fully.

Controls which don't support UIA are duplicated in the accessibility tree:
1 - the owning ToolStripControlHost accessible object as an item of a ToolStrip,
2 - native control accessible object.
They will be fixed in the next commit.